project/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
aiohttp==3.8.5
pyairtable==2.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
google-generativeai==0.3.2
python-multipart==0.0.6
ics==0.7.2
pytz==2023.3
cachetools==5.3.1
pydantic==1.10.12